+++ javascript/modules/client/src/internal/encoding/client.ts
@@ -4,12 +4,148 @@ import { BigNumber } from "@ethersproject/bignumber";
 import { AddressZero } from "@ethersproject/constants";
 import { AVAILABLE_CLIENT_FUNCTION_SIGNATURES } from "../constants/encoding";
 import { FunctionFragment, Interface, Result } from "@ethersproject/abi";
-import { IWithdrawParams } from "../interfaces/client";
+import {
+  IFreezePermissionDecodedParams,
+  IFreezePermissionParams,
+  IGrantPermissionDecodedParams,
+  IGrantPermissionParams,
+  IRevokePermissionDecodedParams,
+  IRevokePermissionParams,
+  IWithdrawParams,
+  PermissionIds,
+} from "../interfaces/client";
+import { keccak256 } from "@ethersproject/keccak256";
+import { toUtf8Bytes } from "@ethersproject/strings";
 
 export function getFunctionFragment(data: Uint8Array): FunctionFragment {
   const hexBytes = bytesToHex(data, true);
   const inter = new Interface(AVAILABLE_CLIENT_FUNCTION_SIGNATURES);
   return inter.getFunction(hexBytes.substring(0, 10));
+}
+export function encodeFreezeAction(
+  params: IFreezePermissionParams,
+): Uint8Array {
+  const daoInterface = DAO__factory.createInterface();
+  const args = unwrapFreezeParams(params);
+  // get hex bytes
+  const hexBytes = daoInterface.encodeFunctionData("freeze", args);
+  return hexToBytes(strip0x(hexBytes));
+}
+
+export function decodeFreezeActionData(
+  data: Uint8Array,
+): IFreezePermissionDecodedParams {
+  const daoInterface = DAO__factory.createInterface();
+  const hexBytes = bytesToHex(data, true);
+  const receivedFunction = daoInterface.getFunction(
+    hexBytes.substring(0, 10) as any,
+  );
+  const expectedFunction = daoInterface.getFunction("freeze");
+  if (receivedFunction.name !== expectedFunction.name) {
+    throw new Error(
+      "The received action is different from the expected one",
+    );
+  }
+  const result = daoInterface.decodeFunctionData("freeze", data);
+  return wrapFreezeParams(result);
+}
+
+function unwrapFreezeParams(
+  params: IFreezePermissionParams,
+): [string, string] {
+  return [
+    params.where,
+    keccak256(toUtf8Bytes(params.permission)),
+  ];
+}
+function wrapFreezeParams(
+  result: Result,
+): IFreezePermissionDecodedParams {
+  return {
+    where: result[0],
+    permissionId: result[1],
+    permission:
+      Object.keys(PermissionIds).find((k) => PermissionIds[k] === result[1])
+        ?.replace(/_ID$/, "") || "",
+  };
+}
+
+export function encodeGrantActionData(
+  params: IGrantPermissionParams,
+): Uint8Array {
+  const daoInterface = DAO__factory.createInterface();
+  const args = unwrapPermissionParams(params);
+  // get hex bytes
+  const hexBytes = daoInterface.encodeFunctionData("grant", args);
+  return hexToBytes(strip0x(hexBytes));
+}
+
+export function decodeGrantActionData(
+  data: Uint8Array,
+): IGrantPermissionDecodedParams {
+  const daoInterface = DAO__factory.createInterface();
+  const hexBytes = bytesToHex(data, true);
+  const receivedFunction = daoInterface.getFunction(
+    hexBytes.substring(0, 10) as any,
+  );
+  const expectedFunction = daoInterface.getFunction("grant");
+  if (receivedFunction.name !== expectedFunction.name) {
+    throw new Error(
+      "The received action is different from the expected one",
+    );
+  }
+  const result = daoInterface.decodeFunctionData("grant", data);
+  return wrapPermissionParams(result);
+}
+
+export function encodeRevokeActionData(
+  params: IRevokePermissionParams,
+): Uint8Array {
+  const daoInterface = DAO__factory.createInterface();
+  const args = unwrapPermissionParams(params);
+  // get hex bytes
+  const hexBytes = daoInterface.encodeFunctionData("revoke", args);
+  return hexToBytes(strip0x(hexBytes));
+}
+
+export function decodeRevokeActionData(
+  data: Uint8Array,
+): IRevokePermissionDecodedParams {
+  const daoInterface = DAO__factory.createInterface();
+  const hexBytes = bytesToHex(data, true);
+  const receivedFunction = daoInterface.getFunction(
+    hexBytes.substring(0, 10) as any,
+  );
+  const expectedFunction = daoInterface.getFunction("revoke");
+  if (receivedFunction.name !== expectedFunction.name) {
+    throw new Error(
+      "The received action is different from the expected one",
+    );
+  }
+  const result = daoInterface.decodeFunctionData("revoke", data);
+  return wrapPermissionParams(result);
+}
+
+function unwrapPermissionParams(
+  params: IGrantPermissionParams | IRevokePermissionParams,
+): [string, string, string] {
+  return [
+    params.where,
+    params.who,
+    keccak256(toUtf8Bytes(params.permission)),
+  ];
+}
+function wrapPermissionParams(
+  result: Result,
+): IGrantPermissionDecodedParams | IRevokePermissionDecodedParams {
+  return {
+    where: result[0],
+    who: result[1],
+    permissionId: result[2],
+    permission:
+      Object.keys(PermissionIds).find((k) => PermissionIds[k] === result[2])
+        ?.replace(/_ID$/, "") || "",
+  };
 }
 
 export function encodeWithdrawActionData(params: IWithdrawParams): Uint8Array {