+++ javascript/modules/client/src/internal/interfaces/client.ts
@@ -9,7 +9,8 @@ import {
   IPagination,
   IPluginInstallItem,
 } from "./common";
-
+import { keccak256 } from "@ethersproject/keccak256";
+import { toUtf8Bytes } from "@ethersproject/strings";
 /** Defines the shape of the general purpose Client class */
 export interface IClient extends IClientCore {
   methods: {
@@ -38,6 +39,18 @@ export interface IClient extends IClientCore {
   };
   encoding: {
     /** Computes the withdraw action payload */
+    grantAction: (
+      daoAddress: string,
+      params: IGrantPermissionParams,
+    ) => DaoAction;
+    revokeAction: (
+      daoAddress: string,
+      params: IRevokePermissionParams,
+    ) => DaoAction;
+    freezeAction: (
+      daoAddress: string,
+      params: IFreezePermissionParams,
+    ) => DaoAction;
     withdrawAction: (
       daoAddresOrEns: string,
       params: IWithdrawParams,
@@ -48,6 +61,9 @@ export interface IClient extends IClientCore {
     ) => Promise<DaoAction>;
   };
   decoding: {
+    grantAction: (data: Uint8Array) => IGrantPermissionDecodedParams;
+    revokeAction: (data: Uint8Array) => IRevokePermissionParams;
+    freezeAction: (data: Uint8Array) => IFreezePermissionDecodedParams;
     withdrawAction: (data: Uint8Array) => IWithdrawParams;
     updateMetadataRawAction: (data: Uint8Array) => string;
     updateMetadataAction: (data: Uint8Array) => Promise<IMetadata>;
@@ -82,6 +98,55 @@ export interface IWithdrawParams {
   tokenAddress?: string;
   reference?: string;
 }
+interface IPermissionParamsBase {
+  where: string;
+  who: string;
+  permission: string;
+}
+interface IPermissionDecodedParamsBase extends IPermissionParamsBase {
+  permissionId: string;
+}
+export interface IGrantPermissionParams extends IPermissionParamsBase {}
+export interface IRevokePermissionParams extends IPermissionParamsBase {}
+export interface IGrantPermissionDecodedParams
+  extends IPermissionDecodedParamsBase {}
+export interface IRevokePermissionDecodedParams
+  extends IPermissionDecodedParamsBase {}
+export interface IFreezePermissionParams {
+  where: string;
+  permission: string;
+}
+export interface IFreezePermissionDecodedParams
+  extends IFreezePermissionParams {
+  permissionId: string;
+}
+
+const Permissions = {
+  UPGRADE_PERMISSION: "UPGRADE_PERMISSION",
+  SET_METADATA_PERMISSION: "SET_METADATA_PERMISSION",
+  EXECUTE_PERMISSION: "EXECUTE_PERMISSION",
+  WITHDRAW_PERMISSION: "WITHDRAW_PERMISSION",
+  SET_SIGNATURE_VALIDATOR_PERMISSION: "SET_SIGNATURE_VALIDATOR_PERMISSION",
+  SET_TRUSTED_FORWARDER_PERMISSION: "SET_TRUSTED_FORWARDER_PERMISSION",
+  ROOT_PERMISSION: "ROOT_PERMISSION",
+  CREATE_VERSION_PERMISSION: "CREATE_VERSION_PERMISSION",
+  REGISTER_PERMISSION: "REGISTER_PERMISSION",
+  REGISTER_DAO_PERMISSION: "REGISTER_DAO_PERMISSION",
+  REGISTER_ENS_SUBDOMAIN_PERMISSION: "REGISTER_ENS_SUBDOMAIN_PERMISSION",
+  MINT_PERMISSION: "MINT_PERMISSION",
+  MERKLE_MINT_PERMISSION: "MERKLE_MINT_PERMISSION",
+  MODIFY_ALLOWLIST_PERMISSION: "MODIFY_ALLOWLIST_PERMISSION",
+  SET_CONFIGURATION_PERMISSION: "SET_CONFIGURATION_PERMISSION",
+};
+
+const PermissionIds = Object.entries(Permissions).reduce(
+  (acc, [k, v]) => ({ ...acc, [k + "_ID"]: keccak256(toUtf8Bytes(v)) }),
+  {} as { [k: string]: string },
+);
+Object.freeze(Permissions);
+export { Permissions };
+Object.freeze(PermissionIds);
+export { PermissionIds };
 
 export enum DaoCreationSteps {
   CREATING = "creating",